+++ app/src/main/cpp/scale.h
@@ -1,25 +1,25 @@
 #ifndef SCALE_H
 #define SCALE_H
 
-double lerp( double t, double min, double max);
-double unlerp( double min, double max, double x);
-double clamp(double v, double min, double max);
+float lerp( float t, float min, float max);
+float unlerp( float min, float max, float x);
+float clamp(float v, float min, float max);
 
 class Scale {
     bool bLogarithmic = true;
 
     //log stuff
-    double a,b, inva, invb;
+    float a,b, inva, invb;
 
     // lin stuff
-    double maxIdx, minFreq, maxFreq;
+    float maxIdx, minFreq, maxFreq;
 
 public:
     void setLogarithmic(bool bLogarithmic_) { bLogarithmic = bLogarithmic_;}
     bool getLogarithmic() const { return bLogarithmic; }
-    void init(double maxIdx,double minFreq, double maxFreq);
-    double forward(double x) const;
-    double backward(double v) const;
+    void init(float maxIdx,float minFreq, float maxFreq);
+    float forward(float x) const;
+    float backward(float v) const;
 };
